requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
redis>=4.0.0
celery>=5.2.0
Pillow>=9.0.0
python-decouple>=3.6
gunicorn>=20.1.0
whitenoise>=6.0.0
django-redis>=5.0.0
django-mptt>=0.14.0
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
markdown>=3.4.0
bleach>=6.0.0
requests>=2.28.0
